Move controller api into its own package.
