+++ api/controller.go
@@ -1,0 +1,11 @@
+package api
+
+// Controller interface for all ingress controllers
+type Controller interface {
+	// Run the controller, returning immediately after it starts or an error occurs.
+	Start() error
+	// Stop the controller, blocking until it stops or an error occurs.
+	Stop() error
+	// Healthy returns true for a healthy controller, false for unhealthy.
+	Health() error
+}
